# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Covalence API",
        "description": "Hybrid GraphRAG knowledge engine API",
        "license": {
            "name": "MIT",
            "identifier": "MIT"
        },
        "version": "0.1.0"
    },
    "paths": {
...
        "/admin/publish/{source_id}": {
            "post": {
                "tags": [
                    "admin"
                ],
                "summary": "Promote a source to a higher clearance level for federation.",
                "operationId": "publish_source",
                "parameters": [
                    {
                        "name": "source_id",
                        "in": "path",
                        "description": "Source ID to publish",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Source published",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PublishResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/audit": {
            "get": {
                "tags": [
                    "admin"
                ],
                "summary": "List recent audit log entries.",
                "operationId": "audit_log",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Maximum number of results (default 20).",
                        "required": false,
                        "schema": {
                            "type": [
                                "integer",
                                "null"
                            ],
                            "format": "int64"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Offset for pagination (default 0).",
                        "required": false,
                        "schema": {
                            "type": [
                                "integer",
                                "null"
                            ],
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Audit log entries",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AuditLogResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AuditLogResponse": {
                "type": "object",
                "description": "Response for an audit log entry.",
                "required": [
                    "id",
                    "action",
                    "actor",
                    "created_at"
                ],
                "properties": {
                    "action": {
                        "type": "string"
                    },
                    "actor": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "target_id": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "uuid"
                    },
                    "target_type": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                }
            },
...
            "PublishResponse": {
                "type": "object",
                "description": "Response for publish operation.",
                "required": [
                    "published"
                ],
                "properties": {
                    "published": {
                        "type": "boolean"
                    }
                }
            },
...
        }
    },
...
}